New Feature: Storage Drop Folder
if btcpay_storagedropfolder is set to a directory path, any file moved or created there will be uploaded to the configured storage provider and then removed from the drop folder. If there is no configured provider, it will wait until one is configured. If you put files in the folder when btcpay is not running, it will queue all files on startup