Maintain invoice list params on navigation
This addresses and potentially closes #1498. As discussed in the issue this addition maintains the list route params on navigation to the invoice details view. It appends the params to the Invoices item in the top navigation so that one can navigate back to the previous list view.

To be discussed: As the params are not namespaced and I'm grabbing them from the `HttpContext.Request.Query` they are also set on the Payment Requests list view.