Update BTCPayServer.csproj
this solves the Visual Studio Bug Problem in a Mac <Content Remove="Properties\launchSettings.json" /> 
plus EnvironmentVariables like Variable name=
could you check this please for me @britttttk on a Mac and confirm this is correct and not introduced any errors or bugs. @NicolasDorier 
this is a fresh copy of Mac install with VS Studio, and BTCPay Server running in my environment , with no problems.
happy if this can solve that old Mac VS bug....  in launchSettings.json & which @pavlenex @rockstardev could reproduce this old error.