Streamline bundles
Optimizes the asset bundles by separating out vendor modules. This allows for more efficient caching and reuse of modules that are only changing infrequently.

As mmost bundles only included a minor part of custom code (which could also be eliminated in the case of the rich text editor configuration), we can also eliminate some bundles alltogether. There might be more optimizations possible, but this is a big change and cleanup already. I'd like to do a follow-up to this one doing those and also updating some of the vendor modules.