Remove option to generate a watch-only wallet
This was sparked by a discussion with @pavlenex: Unfortunately users seem to not store their seed phrase, even though the wallet setup process asks them to do so. This is especially bad for watch-only wallets, as without having it imported into an external wallet, there will be no way to access the funds later on.

Our idea was to remove the option of generating watch-only wallets, which would require the user to first create it another way (e.g. external wallet software or hardware wallet). This leaves the user with the option to import an existing wallet as watch-only, so that they have at least the previous setup instead of relying on them storing the seed properly.