Store branding: Add complementing text and accent colors
Currently we use the current theme text color, independent of the brand color: This is problematic with rather light or dark branding colors, because we might show light text on a light background. This fixes that circumstance by calculating the complementing text color and also adds a slight adjustment for the hover/focus states. The accent color is calculated by decreasing the brightness a bit, which works for light and dark brand colors.