Dynamic Reports
A bit advanced so may put it in a plugin instead. This allows a server admin to create custom reports, which are saved and appear as normal resports in the Reporting features. These reports are only available to Server admins by default but can be marked viewable by non-admins too. This feature only works on Postgres deployments. Security-wise, all queries are run in a transaction and rolled back every time to prevent queries that modify or corrupt. Filters on the rpoerting page can be applied within the sql as parametrs (such as `@StoreId`, `@From`, and `@To`)

![](https://i.imgur.com/SHi2de4.gif)