Allow scheduling installs/updates of future plugins
Sometimes we have to release a BTCPay version that will break all current plugins, and the ony way to fix is to release a new plugin version that nly allows btcpay to be the new version onwards. This means one cannot update plugins to a future version to avoid plugins from breaking. This PR introduces mechanics so that plugin installs and updates can be queued even if the dependencies are not met, so that once they are, they will be installed on startup. If you queue a plugin, and a new update is available, you will be able to replace that with the new version