App: Data additions
Schema additions for App-related functionality. The basis of this PR is to add a new table which acts as a key-value store against a key combination of user/key/version.

You will see that there is a package added: Laraue.EfCoreTriggers.PostgreSql. This package assists in having schema and migrations generated around SQL triggers, in line with EF Core styles. In this case, the trigger created is used to ensure that only the latest version of a value per pey/user is persisted.